+++ nyaa/templates/reports.html
@@ -19,14 +19,14 @@
 			<tr class="reports-row">
 				<td>{{ report.id }}</td>
 				<td>
-					<a href="{{ url_for('view_user', user_name=report.user.username) }}">{{ report.user.username }}</a>
+					<a href="{{ url_for('users.view_user', user_name=report.user.username) }}">{{ report.user.username }}</a>
 					{% if report.user.is_trusted %}
 					<span class="label label-success pull-right">Trusted</span>
 					{% endif %}
 				</td>
 				<td>
 					<a href="{{ url_for('view_torrent', torrent_id=report.torrent.id) }}">{{ report.torrent.display_name }}</a>
-					by <a href="{{ url_for('view_user', user_name=report.torrent.user.username) }}">
+					by <a href="{{ url_for('users.view_user', user_name=report.torrent.user.username) }}">
 					{{ report.torrent.user.username }}</a>
 					{% if g.user.is_superadmin and report.torrent.uploader_ip %}
 					({{ report.torrent.uploader_ip_string }})
