+++ nyaa/templates/view.html
@@ -29,7 +29,7 @@
 		<div class="row">
 			<div class="col-md-1">Submitter:</div>
 			<div class="col-md-5">
-				{% set user_url = torrent.user and url_for('view_user', user_name=torrent.user.username) %}
+				{% set user_url = torrent.user and url_for('users.view_user', user_name=torrent.user.username) %}
 				{%- if not torrent.anonymous and torrent.user -%}
 				<a class="text-{{ torrent.user.userlevel_color }}" href="{{ user_url }}">{{ torrent.user.username }}</a>
 				{%- else -%}
@@ -145,7 +145,7 @@
 		<div class="panel-body">
 			<div class="col-md-2">
 				<p>
-					<a class="text-{{ comment.user.userlevel_color }}" href="{{ url_for('view_user', user_name=comment.user.username) }}">{{ comment.user.username }}</a>
+					<a class="text-{{ comment.user.userlevel_color }}" href="{{ url_for('users.view_user', user_name=comment.user.username) }}">{{ comment.user.username }}</a>
 					{% if comment.user.id == torrent.uploader_id and not torrent.anonymous %}
 					(uploader)
 					{% endif %}
