+++ nyaa/views/__init__.py
@@ -2,8 +2,10 @@ from nyaa.views import (
     account,
     admin,
     site,
+    users,
 )
 
 account_bp = account.bp
 admin_bp = admin.bp
 site_bp = site.bp
+users_bp = users.bp
