+++ nyaa/views/admin.py
@@ -46,19 +46,19 @@ def view_reports():
                 log = log.format(report_id, 'Deleted', torrent_id,
                                  flask.url_for('view_torrent', torrent_id=torrent_id),
                                  report_user.username,
-                                 flask.url_for('view_user', user_name=report_user.username))
+                                 flask.url_for('users.view_user', user_name=report_user.username))
             elif action == 'hide':
                 log = log.format(report_id, 'Hid', torrent_id,
                                  flask.url_for('view_torrent', torrent_id=torrent_id),
                                  report_user.username,
-                                 flask.url_for('view_user', user_name=report_user.username))
+                                 flask.url_for('users.view_user', user_name=report_user.username))
                 torrent.hidden = True
                 report.status = 1
             else:
                 log = log.format(report_id, 'Closed', torrent_id,
                                  flask.url_for('view_torrent', torrent_id=torrent_id),
                                  report_user.username,
-                                 flask.url_for('view_user', user_name=report_user.username))
+                                 flask.url_for('users.view_user', user_name=report_user.username))
                 report.status = 2
 
             adminlog = models.AdminLog(log=log, admin_id=flask.g.user.id)
